Added polyfills. Fix server build
