+++ .babelrc
@@ -4,15 +4,14 @@
     [
       "@babel/env", {
         "targets": {
-          "node": "10.7",
           "browsers": "> 1%, not dead"
         },
         "exclude": ["transform-regenerator", "transform-async-to-generator"],
-        "debug": true
+        "debug": true,
+        "modules": false,
+        "useBuiltIns": "usage"
       }
     ]
   ],
-  "plugins": [
-    "async-to-promises"
-  ]
+  "plugins": ["async-to-promises"]
 }
