+++ src/web/components/SlackButton.jsx
@@ -12,7 +12,7 @@ class SlackButton extends Component {
           height={height}
           width={width}
           src="https://platform.slack-edge.com/img/add_to_slack.png"
-          srcset="https://platform.slack-edge.com/img/add_to_slack.png 1x, https://platform.slack-edge.com/img/add_to_slack@2x.png 2x"
+          srcSet="https://platform.slack-edge.com/img/add_to_slack.png 1x, https://platform.slack-edge.com/img/add_to_slack@2x.png 2x"
         />
       </a>
     );
